feat(observability.logs): update logs unsubscribe modal
resolves: #MAOBS-235
