+++ packages/manager/modules/logs-to-customer/public/translations/logSubscription/Messages_fr_FR.json
@@ -8,9 +8,10 @@
   "log_subscription_empty_tile_button_subscribe": "S'abonner",
   "log_subscription_list_title": "Utiliser vos logs",
   "log_subscription_list_button_subscribe_more": "S'abonner sur un autre flux de données",
-  "log_subscription_terminate_modal_headline": "Se désabonner",
-  "log_subscription_terminate_modal_delete_input_label": "Veuillez entrer le mot « TERMINATE » pour vous désabonner du flux de données.",
-  "log_subscription_terminate_modal_confirm_button_label": "Se désabonner",
+  "log_subscription_terminate_modal_headline": "Se désabonner ?",
+  "log_subscription_terminate_modal_description": "Se désabonner au flux de données pour cette ressource entraînera l'arrêt définitif de la collecte et la suppression des données associées.",
+  "log_subscription_terminate_modal_cancel_button": "Non, annuler",
+  "log_subscription_terminate_modal_confirm_button": "Oui, se désabonner",
   "log_subscription_unsubscribe_success": "Votre désabonnement a réussi.",
   "log_subscription_unsubscribe_error": "Votre désabonnement a rencontré une erreur. {{error}}",
   "log_subscription_subscribe_success": "Votre abonnement a réussi.",
