+++ packages/manager/modules/logs-to-customer/public/translations/logSubscription/Messages_it_IT.json
@@ -8,11 +8,14 @@
   "log_subscription_empty_tile_button_subscribe": "Abbonati",
   "log_subscription_list_title": "Utilizzare i log",
   "log_subscription_list_button_subscribe_more": "Abbonati a un altro flusso di dati",
-  "log_subscription_terminate_modal_headline": "Disattivare l'abbonamento",
+  "log_subscription_terminate_modal_headline": "Vuoi disiscriverti?",
   "log_subscription_terminate_modal_delete_input_label": "Inserisci la parola \"TERMINATE\" per disiscriverti dal flusso di dati.",
   "log_subscription_terminate_modal_confirm_button_label": "Disattivare l'abbonamento",
   "log_subscription_unsubscribe_success": "La tua disiscrizione è andata a buon fine.",
   "log_subscription_unsubscribe_error": "Si è verificato un errore durante la disiscrizione: {{error}}",
   "log_subscription_subscribe_success": "L'iscrizione è andata a buon fine.",
-  "log_subscription_subscribe_error": "Si è verificato un errore durante l’iscrizione: {{error}}"
+  "log_subscription_subscribe_error": "Si è verificato un errore durante l’iscrizione: {{error}}",
+  "log_subscription_terminate_modal_description": "La disiscrizione dal flusso dei dati per questa risorsa comporterà l'interruzione definitiva della raccolta e l'eliminazione dei dati associati.",
+  "log_subscription_terminate_modal_cancel_button": "No, annulla",
+  "log_subscription_terminate_modal_confirm_button": "Sì, voglio disiscrivermi"
 }
